requirements.txt
fastapi
uvicorn
pymongo
cryptography
qdrant-client
transformers
torch
peft
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv
sentence-transformers
pytest
httpx